Import some code from Ruby itself to be able to modify it
By importing `rb_profile_frames` we can modify it to fix the issue with
wrong labels for blocks.

As documented in the code, I don't yet understand enough of the Ruby VM
to understand if I'm doing something wrong OR if the original
implementation has a bug.

I'm somewhat leaning towards a bug in the original implementation since
it doesn't seem to be used very often but maybe I'm not giving the Ruby
devs enough credit and there's a good reason for that.