Correctly skip dummy frames at the bottom of the stack
The previous hack of "always skip one frame at the bottom of the stack"
did not work correctly in all cases -- when sampling a thread other than
`Thread.main`, the code from `rb_profile_frames` already skips the
dummy frame while iterating (hint: it's the branch where `i` doesn't
get incremented), whereas not for the `Thread.main`.

I've borrowed the approach from MRI's `backtrace_each`, but as usual
I'm curious if this behavior of `rb_profile_frames` is a
bug/oversight, or if I'm missing something.