Fix builds on windows by refactoring usage of RUBY_MJIT_HEADER
Windows builds on CI were failing with
"multiple definition of `something...'" (such as `rb_vm_ep_local_ep')
and it turns out that (some?) Windows versions of Ruby define a bunch of
functions in the RUBY_MJIT_HEADER itself without marking them as
"static" (e.g. not visible to the outside of the file), and thus the
linker then complains when linking together several files which all
have these non-private symbols.

To solve this I've decided to centralize all usage of this header on
a single file -- `ruby_shards.c`.

This file combines the old `ruby_3.0.0.c` and `cfunc_frames_backport.c`
files, and thus solves the issue entirely.