Introduce struct raw_location to better represent the data from rb_profile_frames
This allows us to stop adding more and more random arrays to
`rb_profile_frames` and instead return a better structured "object"
with a few useful flags that avoids needing to "guess" from the
caller-side.

This can later lead to a few more simplifications (as the code is
differing more and more from the original code that depended on the
original `rb_profile_frames`, but for now most of the structure is
unchanged.

This whole change will also enable us to start exposing the name of the
called method (via the called_id inside the callable_method_entry).