Add initial support for Ruby 2.5!
This works not by using the MJIT headers (which do not exist), but by
piggybacking on the actual Thread#backtrace_locations API, breaking
into their private APIs and getting (some of) what we get using the
MJIT headers on newer Rubies.

It's not perfect, but it's a first start! Really cool too.

The code is somewhat messy, will probably clean up soon :)