Add proper qualified_method_name for refinements
I... now know more about refinements than I really wanted to. They're
really bizarre. A refinement is a set of two modules:

```ruby
module Foo
  refine Bar do
    # ...
  end
end
```

There are two modules here: `Foo`, and `#<refinement:Bar@Foo>`.

The weird thing is that actually because the return value of a module
definition is the last value, the `#<refinement:Bar@Foo>` is actually
the return value of the definition above, and thus this module is
actually visible to regular Ruby code!

But if you lose that reference, you won't find it again. MRI stores
the reference inside an instance variable of `Foo`, with a name that
is not accessible from Ruby code.

Not sure why, actually, since it could just be a regular constant...
¯\_(ツ)_/¯ .

Anyway, I've replicated a part of `Module#to_s` and used it to implement
the desired `#qualified_method_name` format for refinements.