Generalize addition of {block} suffix in qualified_method_name
Blocks (or evals, which for our purposes are just a really weird block)
can be inside many things; I've realized the "whack-a-mole" approach
to discovering more cases where the {block} suffix needed to be added
was incorrect, and it should be thought as something ortogonal to
the other queries and checks while building the `qualified_method_name`.

This also solves one corner case for `eval` which wasn't properly
being handled (and in the next commit one other corner case will also
rely on this).