Re-add support for Ruby 2.5, using the debase-ruby_core_source
I had previously tried to copy-paste bits and pieces from the Ruby VM
header files, but had come up empty, and thus I decided to support
older Rubies by piggybacking on the Thread#backtrace_locations API.

This had a lot of issues -- introduced a lot of alternative code paths
(that I tended to forget existed from time to time), as well as
limiting some of the data that could be gathered.

Instead, this commit rebuilds the 2.5 support on top of the
`debase-ruby_core_source` gem, which basically ships a copy of
internal VM headers for released Ruby versions. Thus, it's a lot
closer to what we get on Ruby 2.6+ from the MJIT header, and the
small diff needed to add support for Ruby 2.5 on top of of it
proves that.

Nice!