Remove pending from test that is now fixed
This got fixed in the previous commit, it turns out that the
`main_object_instance = backtracie_rb_vm_top_self();` was broken on
Ruby 2.5 (and did not work for 2.4 at all).

It got replaced with `TOPLEVEL_BINDING.eval("self")` which works on all
Rubies (even pre-2.3, which we don't support), which is even better
because no more poking at VM insides to get this.