Normalise the code for finding paths for cfuncs
The C API now does no kind of traversal of previous frames. Instead, the
Ruby API is in charge of synthesising a path for cframes.

If there _is_ a higher-level ruby frame for a cfunc, then we use the
path from that, emulating the behaviour of Thread#backtrace_locations

If there is _not_ a higher-level ruby frame, then we return the string
"(in native code)" and also set a flag path_is_synthetic on the
location.