Fix accessing instance variables in Ruby 3.2
`mod_to_s_singleton` is covered by the
"when sampling a method defined in a module's singleton class" spec.

In Ruby 3.2, due to the introduction of object shapes, our existing
code to directly read the `__attached__` instance variable stopped
working.

To fix this, let's use the regular Ruby API (`rb_ivar_get`) to read
instance variables.

Looking at the implementation for legacy (pre-3.2) Rubies, it doesn't
do much more work than our direct path, so let's go with this for
now and later if it turns out to be a performance hotspot we can
re-add a fast path.