Correctly get name of included module in Ruby 2.3 and 3.2
I wasn't quite able to figure out what changed in Ruby 3.2, but
the "when sampling an eval" started failing in exactly the same way it
was failing for 2.3 (yes, the symmetry is fun).

...and it turns out the same fix works for both, which is to handle
the T_ICLASS in a special way.

I will admit I'm not entirely confident about this change, but we do
have test coverage showing that I didn't break anything else, so I
guess we'll find out if this is enough or if I missed any corner case.