Attempt to disable Ruby 3.2 and above for Windows
Upstream dropped support for MJIT on Windows in 3.2
(https://bugs.ruby-lang.org/issues/18824) which means the mjit header
is no longer available.

In the future, we may choose to support Windows using
`debase-ruby_core_source` but to be honest I don't think there's a lot
of interest for Ruby on Windows so we should probably wait until
someone actually cares enough to request or contribute it.